BitMap and BloomFilter

+++ 【数据结构】哈希之位图与布隆过滤器相关操作/BitMap.h
@@ -15,9 +15,9 @@ typedef struct BitMapNode
 void BitMapInit(BitMap* pbm,size_t len);
 void BitMapDestroy(BitMap* pbm);
 
-void BitMapSet(BitMap* pbm,size_t x);
-void BitMapReset(BitMap* pbm,size_t x);
-int BitMapTest(BitMap* pbm,size_t x);
+void BitMapSet(BitMap* pbm, int x);
+void BitMapReset(BitMap* pbm, int x);
+int BitMapTest(BitMap* pbm, int x);
 void TestBitMap();
